refactor: use courses instead of cursos
