+++ frontend/src/pages/dashboard/layoutConfig.js
@@ -2,8 +2,8 @@ export function alumnosNavConfig() {
   return {
     links: [
       { path: '.', name: 'Inicio' },
-      { path: 'cursos', name: 'Cursos' },
-      { path: 'mentores', name: 'Mentores' },
+      { path: 'courses', name: 'Cursos' },
+      { path: 'mentors', name: 'Mentores' },
       { path: 'actividades', name: 'Actividades' },
       { path: 'eventos', name: 'Eventos' }
     ],
